Cleaned up builtin listener matchers
