--- slack_bolt/middleware/function_listener_matches/__init__.py
@@ -1,5 +1,0 @@
-from .function_listener_token import FunctionListenerToken
-
-__all__ = [
-    "FunctionListenerToken",
-]
