+++ slack_bolt/middleware/function_token/__init__.py
@@ -1,0 +1,5 @@
+from .function_listener_token import FunctionToken
+
+__all__ = [
+    "FunctionToken",
+]
